libc: a function declared by two headers has one type in both
Several functions have a declaration in more than one bundled header, or
one the standard fixes, and the declarations disagreed:

- `popen` and `fdopen` took `char *` where POSIX declares `const char *`;
- on Windows, <stdio.h> declared `wcslen`, `wcscmp`, `wcschr` and their
  neighbours with `long long` for `size_t`, `int` for `wchar_t` and no
  `const`, where <wchar.h> declares the standard's types; <wchar.h> in
  turn gave `wcslen` an `unsigned long long` result where `size_t` is
  `unsigned long` on the LP64 targets;
- `_read`, `_write` and `_lseek` in <stdio.h> took and returned
  `long long` where <io.h> and the CRT use `int`, `unsigned int` and
  `long`, and `_byteswap_ulong` used `unsigned int` where <intrin.h> uses
  `unsigned long`;
- `recvfrom`, `sendto` (<sys/socket.h>) and `gethostname` (<unistd.h>)
  had the POSIX shape on Windows, where <winsock2.h> declares Winsock's;
- `ConvertStringSecurityDescriptorToSecurityDescriptorW` in <windows.h>
  took `void *` and `LPDWORD` where <sddl.h> and the SDK take
  `PSECURITY_DESCRIPTOR *` and `PULONG`.

Each now carries the standard's, the CRT's or the SDK's type in every
header that declares it, so a unit including several of them sees one type
per function.

+++ libc/include/wchar.h
@@ -168,7 +168,7 @@ static inline wchar_t *wmemset(wchar_t *d, wchar_t c, size_t n) {
 #endif
 
 // Wide-string handling (C99 7.24.4). `size_t` comes from <stddef.h>.
-unsigned long long wcslen(const wchar_t *s);
+size_t wcslen(const wchar_t *s);
 int wcscmp(const wchar_t *s1, const wchar_t *s2);
 int wcsncmp(const wchar_t *s1, const wchar_t *s2, size_t n);
 wchar_t *wcschr(const wchar_t *s, wchar_t c);
